fix color panel deps.
